+++ test/traits/logical.cpp
@@ -66,7 +66,7 @@ namespace zacc { namespace test {
             auto actual = !condition;
             auto data = actual.data();
 
-            for(int i = 0; i < size_v<TypeParam>; i++)
+            for(size_t i = 0; i < size_v<TypeParam>; i++)
             {
                 VASSERT_EQ(data[i], expected[i]);
             }
@@ -96,7 +96,7 @@ namespace zacc { namespace test {
             auto actual = !condition;
             auto data = actual.data();
 
-            for(int i = 0; i < size_v<TypeParam>; i++)
+            for(size_t i = 0; i < size_v<TypeParam>; i++)
             {
                 VASSERT_EQ(data[i], expected[i]);
             }
@@ -125,7 +125,7 @@ namespace zacc { namespace test {
             auto actual = !condition;
             auto data = actual.data();
 
-            for(int i = 0; i < size_v<TypeParam>; i++)
+            for(size_t i = 0; i < size_v<TypeParam>; i++)
             {
                 VASSERT_EQ(data[i], expected[i]);
             }
@@ -154,7 +154,7 @@ namespace zacc { namespace test {
             auto actual = !condition;
             auto data = actual.data();
 
-            for(int i = 0; i < size_v<TypeParam>; i++)
+            for(size_t i = 0; i < size_v<TypeParam>; i++)
             {
                 VASSERT_EQ(data[i], expected[i]);
             }
@@ -184,7 +184,7 @@ namespace zacc { namespace test {
             auto data = actual.data();
 
 
-            for(int i = 0; i < size_v<TypeParam>; i++)
+            for(size_t i = 0; i < size_v<TypeParam>; i++)
             {
                 VASSERT_EQ(data[i], expected[i]);
             }
